+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## [Unreleased]
+* allow exclude crown dependency from correspondent registry
 * update rubocop configuration
 * bump dependent gem `ndr_dev_support` to v6.0